LiquidTransportParams.h
Added some comments plus members
+       DenseMatrix  visc_Eij;
+       DenseMatrix  visc_Sij;
These members are not populated in any way at this point.  A message
to this effect is written in
TransportFactory::getLiquidTransportData. 


LiquidTransport.h LiquidTransport.cpp

Changed the types of viscosity models from 
   const int LVISC_CONSTANT     = 0;
-  const int LVISC_WILKES       = 1;
-  const int LVISC_MIXTUREAVG   = 2;
to
   const int LVISC_CONSTANT     = 0;
+  const int LVISC_INTERACTION  = 1;
+  const int LVISC_AVG_ENERGIES = 2;
These are not yet implemented.

Added members 
+    DenseMatrix m_visc_Eij;
+    DenseMatrix m_visc_Sij;
+    vector_fp  m_hydrodynamic_radius;
+    vector_fp  m_visc_logA; //logarithm of coefficient

Removed unneeded members:
-    vector_fp   m_eps;
-    doublereal m_sqrt_t;
-    doublereal m_t14;
-    doublereal m_t32;
-    doublereal m_sqrt_kbt;

Changed the name of m_cond to m_condSpecies in agreement with the
m_viscSpecies pattern.

Implemented some code in LiquidTransport::viscosity() to compute the
molecular interactions relevant to mixture viscosity, BUT none of the
required parameters or flags are implemented at this point.



git-svn-id: https://cantera.googlecode.com/svn/cantera18/changes/liquidTransportDevelop@198 02a645c2-efd0-11dd-984d-ab748d24aa7e