Added mapping,m_LTImodelMap, between LiquidTranMixingModel and the model name.
Added new method to build LiquidTranInteraction objects.
    virtual LiquidTranInteraction* newLTI( const XML_Node &trNode, 
					   TransportPropertyList tp_ind, 
					   LiquidTransportParams& trParam) ;
This methods calls teh new operator and the init() method for those
objects.  It also call the setParameters() method for those objects
that require additional trParam access.

In TransportFactory::getLiquidInteractionsTransportData() the majority
of the work is done by a switch statement that fills the different
LiquidTranInteraction* in LiquidTransportParams by calling newLTI().
The switch statement is over the type of transport property
(i.e. viscosity).  After some debugging checks we can remove the rest
of that method. (all of the XML node parsing  is moved to the
LiquidTranInteraction constructor.



git-svn-id: https://cantera.googlecode.com/svn/cantera18/changes/liquidTransportDevelop@271 02a645c2-efd0-11dd-984d-ab748d24aa7e