Added member to TransportBase to allow specification of the reference velocity.  This allows selection of a mass-averaged, mole-averaged or solvent specified reference velocity using the member m_velocityBasis or the methods setVelocityBasis() and getVelocityBasis().  Parsing of input needs to  be added for this still.  An enum has been added   enum VelocityBasis {     VB_MOLEAVG = -2,     VB_MASSAVG = -1  }; Other values can correspond to species indices.
In LiquidTransport, the Stefan Maxwell solve now checks to see what
the m_velocityBasis member says the reference velocity should be and
fills the matrix accordingly.  To allow mass averaged we have added
members m_massfracs and m_massfracs_tran.  

Added methods to extract the diffusion velocity in a similar manner to
the diffusive fluxes.  The diffusion velocity will be needed to not
force the porosity/tortuosity into the Transport classes for
porous flow.  These methods are getSpeciesVdiff() and
getSpeciesVdiffES(). 



git-svn-id: https://cantera.googlecode.com/svn/cantera18/changes/liquidTransportDevelop@273 02a645c2-efd0-11dd-984d-ab748d24aa7e