[Thermo] Adjust NASA polynomials to eliminate discontinuities
We expect the high- and low-temperature NASA polynomials for each species to be
continuous in a number of places, despite many frequently used thermo databases
having data that is inconsistent.

In addition to warning about discontinuities, We now modify the provided
coefficients to ensure continuity at the midpoint temperature. This resolves
numerical issues, e.g. with the algorithm used for setting the state of a
reactor network.

git-svn-id: https://cantera.googlecode.com/svn/cantera/trunk@2238 02a645c2-efd0-11dd-984d-ab748d24aa7e