[1D] Fix save/restore when mechanism size differs
The failures were caused when attempting to restore the tolerance vectors, which
have a value for each species. Since these tolerances are usually the same for
all species, the last value in the array can be used to extend the array to the
required length.

Also add some tests for this feature.

git-svn-id: https://cantera.googlecode.com/svn/cantera/trunk@2480 02a645c2-efd0-11dd-984d-ab748d24aa7e