[SCons] Add configuration option for specifying required Boost libraries.
The set of Boost libraries which are required by Cantera, through the use of
Boost.Thread, depends on the Boost version. The static libraries are needed
when compiling anything that links to Cantera, so these libraries are copied
to the Cantera installation directory on Windows, where compiling Boost from
source is a burden.

git-svn-id: https://cantera.googlecode.com/svn/cantera/branches/2.1@2530 02a645c2-efd0-11dd-984d-ab748d24aa7e