[Matlab] Remove broken ck2ctml function
There are three problems with this function. It does not work, since
it calls ctmethods with the wrong number of arguments. Even if that
were fixed, it wouldn't do what it claims, as the method it calls
generates a CTI file, not an XML file. Finally, it is redundant, as
the 'ck2cti' function can be used for generating the CTI file, and the
CTI to XML conversion is handled automatically when importing the
phase.

Therefore, this function is being removed.

git-svn-id: https://cantera.googlecode.com/svn/cantera/branches/2.1@2581 02a645c2-efd0-11dd-984d-ab748d24aa7e