[SCons] Check for incompabible BLAS/LAPACK when building Matlab toolbox
Matlab's practice of substituting an "8 byte integer" BLAS/LAPACK
library for the "4 byte integer" BLAS/LAPACK libraries that Cantera is
designed to use causes Cantera to crash. This can only be avoided by
statically linking Cantera to BLAS and LAPACK, e.g. by using the
BLAS/LAPACK implementations included with Cantera.

Since the Matlab toolbox itself needs to be a shared library, it is
not usually possible to link statically link it with vendor-provided
BLAS/LAPACK libraries, since they are not compiled with the
appropriate options, i.e. '-fPIC'.

Resolves Issue 34.

git-svn-id: https://cantera.googlecode.com/svn/cantera/trunk@2617 02a645c2-efd0-11dd-984d-ab748d24aa7e