Introduce a more general method of caching results
The ValueCache class is intended to standardize the method for implementing
value caching in Cantera, and to reduce the need to introduce additional member
variables to store cached variables and the thermodynamic state at which the
cached values were computed.

The initial usage is to replace the use of the m_tlast variable in IdealGasPhase
as an example.

git-svn-id: https://cantera.googlecode.com/svn/cantera/trunk@2820 02a645c2-efd0-11dd-984d-ab748d24aa7e