[Cython/Reactor] Print messages from exceptions raised in Python callbacks
Unlike the normal case for CallbackError, where the error message is visible
when it is re-raised in Python, errors occuring during reactor network
integration aren't necessarily terminal, so they can't simply be re-raised after
returning to Python. But we still want to display the information they contain,
so we need to extract that information from the Python exception objects in the
what() function that is called by cvodes_rhs.

git-svn-id: https://cantera.googlecode.com/svn/cantera/trunk@3170 02a645c2-efd0-11dd-984d-ab748d24aa7e