[Thermo] Simplify interface for adding elements to a Phase
Phase::addElements(string name, ...) is the single entry point for adding
elements. It always perform the uniqueness check, and can do the extra
additional work required to add elements to a phase that already has species,
removing the need for "freezing" the phase's elements. Parsing the XML tree for
elements is now handled in "installElements". The variant element-adding methods
are deprecated.

git-svn-id: https://cantera.googlecode.com/svn/cantera/trunk@3193 02a645c2-efd0-11dd-984d-ab748d24aa7e