[Test] Use Cantera shared library for tests, when possible
On Linux systems, link tests against the Cantera shared library, rather than the
static library to reduce the size of the compiled test suite. This saves over 1
GB of disk space when compiling with debug symbols enabled, and also reduces the
run time of the test suite.

This does not work on Windows, where the shared library only exports the Cantera
C interface.

git-svn-id: https://cantera.googlecode.com/svn/cantera/trunk@3199 02a645c2-efd0-11dd-984d-ab748d24aa7e